+++ kolibri/core/content/management/commands/exportchannel.py
@@ -3,7 +3,7 @@ import os
 
 from ...utils import paths
 from ...utils import transfer
-from kolibri.tasks.management.commands.base import AsyncCommand
+from kolibri.core.tasks.management.commands.base import AsyncCommand
 
 logging = logger.getLogger(__name__)
