+++ kolibri/core/tasks/apps.py
@@ -6,7 +6,7 @@ from django.apps import AppConfig
 
 
 class KolibriTasksConfig(AppConfig):
-    name = 'kolibri.tasks'
+    name = 'kolibri.core.tasks'
     label = 'kolibritasks'
     verbose_name = 'Kolibri Tasks'
